refactor: keep OAT normalization within quality limits
